add user schema and onCreateUser adds schedule
