--- imports/server.js/accounts.js
@@ -1,4 +1,0 @@
-Accounts.onCreateUser((options, user) => {
-  console.log('options, user', options, user);
-  return user;
-});
